vm: move opcode implementations to eips.go
Moves the actual opcode implementations to `eips.go`
which contains opcodes added through EIPs. Also add
method docs to the new functions.
